refactor: delete hooks module and remove rewrite_stash feature flag
Inline push_hooks and stash_hooks logic into daemon.rs, delete the
commands/hooks/ module entirely, and remove the now-unused rewrite_stash
feature flag (stash handling is always-on).
