--- src/commands/hooks/mod.rs
@@ -1,2 +1,0 @@
-pub mod push_hooks;
-pub mod stash_hooks;
