+++ src/commands/mod.rs
@@ -13,7 +13,6 @@ pub mod flush_metrics_db;
 pub mod git_ai_handlers;
 pub mod git_handlers;
 pub mod git_hook_handlers;
-pub mod hooks;
 pub mod install;
 pub mod install_hooks;
 pub mod log;
