+++ src/feature_flags.rs
@@ -78,7 +78,6 @@ macro_rules! define_feature_flags {
 // Define all feature flags in one place
 // Format: struct_field: file_and_env_name, debug = <bool>, release = <bool>
 define_feature_flags!(
-    rewrite_stash: rewrite_stash, debug = true, release = true,
     auth_keyring: auth_keyring, debug = false, release = false,
     git_hooks_enabled: git_hooks_enabled, debug = false, release = false,
     git_hooks_externally_managed: git_hooks_externally_managed, debug = false, release = false,
@@ -95,7 +94,7 @@ impl FeatureFlags {
 
     /// Build FeatureFlags from environment variables
     /// Reads from GIT_AI_* prefixed environment variables
-    /// Example: GIT_AI_REWRITE_STASH=true, GIT_AI_AUTH_KEYRING=false
+    /// Example: GIT_AI_AUTH_KEYRING=true
     /// Falls back to defaults for any invalid or missing values
     #[allow(dead_code)]
     pub fn from_env() -> Self {
@@ -132,10 +131,8 @@ mod tests {
     #[test]
     fn test_default_feature_flags() {
         let flags = FeatureFlags::default();
-        // Test that defaults are set correctly based on debug/release mode
         #[cfg(debug_assertions)]
         {
-            assert!(flags.rewrite_stash);
             assert!(!flags.auth_keyring);
             assert!(!flags.git_hooks_enabled);
             assert!(!flags.git_hooks_externally_managed);
@@ -145,7 +142,6 @@ mod tests {
         }
         #[cfg(not(debug_assertions))]
         {
-            assert!(flags.rewrite_stash);
             assert!(!flags.auth_keyring);
             assert!(!flags.git_hooks_enabled);
             assert!(!flags.git_hooks_externally_managed);
@@ -158,28 +154,23 @@ mod tests {
     #[test]
     fn test_from_deserializable() {
         let deserializable = DeserializableFeatureFlags {
-            rewrite_stash: Some(false),
             auth_keyring: Some(true),
             ..Default::default()
         };
 
         let flags = FeatureFlags::from_deserializable(deserializable);
-        assert!(!flags.rewrite_stash);
         assert!(flags.auth_keyring);
     }
 
     #[test]
     #[serial_test::serial]
     fn test_from_env_and_file_defaults_only() {
-        // No file flags, env should be empty
         unsafe {
-            std::env::remove_var("GIT_AI_REWRITE_STASH");
             std::env::remove_var("GIT_AI_AUTH_KEYRING");
         }
 
         let flags = FeatureFlags::from_env_and_file(None);
         let defaults = FeatureFlags::default();
-        assert_eq!(flags.rewrite_stash, defaults.rewrite_stash);
         assert_eq!(flags.auth_keyring, defaults.auth_keyring);
     }
 
@@ -187,25 +178,21 @@ mod tests {
     #[serial_test::serial]
     fn test_from_env_and_file_file_overrides() {
         unsafe {
-            std::env::remove_var("GIT_AI_REWRITE_STASH");
             std::env::remove_var("GIT_AI_AUTH_KEYRING");
         }
 
         let file_flags = DeserializableFeatureFlags {
-            rewrite_stash: Some(true),
             auth_keyring: Some(true),
             ..Default::default()
         };
 
         let flags = FeatureFlags::from_env_and_file(Some(file_flags));
-        assert!(flags.rewrite_stash);
         assert!(flags.auth_keyring);
     }
 
     #[test]
     fn test_serialization() {
         let flags = FeatureFlags {
-            rewrite_stash: true,
             auth_keyring: true,
             git_hooks_enabled: false,
             git_hooks_externally_managed: false,
@@ -215,7 +202,6 @@ mod tests {
         };
 
         let serialized = serde_json::to_string(&flags).unwrap();
-        assert!(serialized.contains("rewrite_stash"));
         assert!(serialized.contains("auth_keyring"));
         assert!(serialized.contains("git_hooks_enabled"));
         assert!(serialized.contains("git_hooks_externally_managed"));
@@ -227,7 +213,6 @@ mod tests {
     #[test]
     fn test_clone_trait() {
         let flags = FeatureFlags {
-            rewrite_stash: true,
             auth_keyring: true,
             git_hooks_enabled: true,
             git_hooks_externally_managed: false,
@@ -236,7 +221,6 @@ mod tests {
             checkpoint_debug_log: true,
         };
         let cloned = flags.clone();
-        assert_eq!(cloned.rewrite_stash, flags.rewrite_stash);
         assert_eq!(cloned.auth_keyring, flags.auth_keyring);
         assert_eq!(cloned.git_hooks_enabled, flags.git_hooks_enabled);
         assert_eq!(
